+++ src/main/java/com/sparta/codesourcecommunity/member/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.sparta.codesourcecommunity.member.controller;
 
 import com.sparta.codesourcecommunity.common.CommonResponseDto;
-import com.sparta.codesourcecommunity.member.JwtUtil;
+import com.sparta.codesourcecommunity.security.JwtUtil;
 import com.sparta.codesourcecommunity.member.dto.MemberRequestDto;
 import com.sparta.codesourcecommunity.member.service.MemberService;
 import jakarta.servlet.http.HttpServletResponse;
