+++ src/main/java/com/sparta/codesourcecommunity/security/JwtAuthorizationFilter.java
@@ -1,8 +1,7 @@
-package com.sparta.codesourcecommunity.member;
+package com.sparta.codesourcecommunity.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.codesourcecommunity.common.CommonResponseDto;
-import com.sparta.codesourcecommunity.security.MemberDetailsService;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -29,14 +28,14 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
     private final ObjectMapper objectMapper;
 
 
-
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
+        FilterChain filterChain)
         throws ServletException, IOException {
         String token = jwtUtil.resolveToken(request);
 
-        if(Objects.nonNull(token)){ //token != null
-            if(jwtUtil.validateToken(token)) {
+        if (Objects.nonNull(token)) { //token != null
+            if (jwtUtil.validateToken(token)) {
                 Claims info = jwtUtil.getUserInfoFromToken(token);
 
                 // 인증정보에 유저정보(email) 넣기
@@ -44,16 +43,18 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
                 //username -> user 조회 -> userDetails 에 담고 -> authentication 의 principal 에 담고
                 SecurityContext context = SecurityContextHolder.createEmptyContext();
                 UserDetails memberDetails = memberDetailsService.getMemberDetails(email);
-                Authentication authentication = new UsernamePasswordAuthenticationToken(memberDetails,null);
+                Authentication authentication = new UsernamePasswordAuthenticationToken(
+                    memberDetails, null);
                 //-> securityContent 에 담고
                 context.setAuthentication(authentication);
                 //-> SecurityContextHolder 에 담고
                 SecurityContextHolder.setContext(context);
                 //-> 이제 @AuthenticationPrincipal 로 조회할 수 있음
 
-            } else{
+            } else {
                 // 인증정보가 존재하지 않을때
-                CommonResponseDto commonResponseDto = new CommonResponseDto("토큰이 유효하지 않습니다.", HttpStatus.BAD_REQUEST.value());
+                CommonResponseDto commonResponseDto = new CommonResponseDto("토큰이 유효하지 않습니다.",
+                    HttpStatus.BAD_REQUEST.value());
                 response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
                 response.setContentType("application/json; charset=UTF-8");
                 response.getWriter().write(objectMapper.writeValueAsString(commonResponseDto));
